Feature/global configuration profiles (#90)
Different source code projects might want to use different component or license metadata. This PR introduces the concept of configuration profiles and allows users to choose one when invoking a build step.

This is a breaking change in the regard that the existing configuration is lost. Therefore, the version has been incremented to 2.0.0